proyecto_bbdd/requirements.txt
scikit-learn==1.4.0
pandas==2.2.1
PyYAML==6.0.1
pyarrow==15.0.0
fastparquet==2024.2.0
jupyterlab==4.1.5



